Move labels to histo_functions.
